refactor(utils): finish _datashader.py extraction, kill back-import (#696)
Move 13 datashader helper symbols verbatim from utils.py into _datashader.py
(_ax_show_and_transform, _compute_datashader_canvas_params,
_get_extent_and_range_for_datashader_canvas, _datashader_canvas_from_dataframe,
_create_image_from_datashader_result, _DS_REDUCTION_FUNCS,
_datashader_aggregate_with_function, _datashader_get_how_kw_for_spread,
_prepare_transformation, _apply_cmap_alpha_to_datashader_result,
_datashader_map_aggregate_to_color, _hex_no_alpha, _convert_alpha_to_datashader_range),
removing the utils back-import that was the repo's only near-cycle. _datashader.py
now imports only downward from utils (_fast_extent, _make_continuous_mappable, to_hex).

Also fixes the _datshader_get_how_kw_for_spread -> _datashader_get_how_kw_for_spread
typo (def + caller). Repoints render.py (5 names) and test imports
(test_utils, test_render_points, test_render_shapes) to _datashader.

No behavior change (verbatim move; all-private symbols). Verified: no import cycle;
195 non-visual tests pass; ruff/format clean on moved code. Pre-existing ruff D205 /
mypy Any-return debt in utils.py (#703/#705) is unrelated and fixed by #714;
--no-verify used for that reason only.
