+++ src/spatialdata_plot/pl/_datashader.py
@@ -8,27 +8,41 @@ from __future__ import annotations
 from copy import copy
 from typing import Any, Literal
 
+import dask
 import dask.dataframe as dd
 import datashader as ds
 import matplotlib
 import matplotlib.colors
+import matplotlib.image
+import matplotlib.transforms as mtransforms
 import numpy as np
+import numpy.ma as ma
+import numpy.typing as npt
 import pandas as pd
+from datashader.core import Canvas
+from geopandas import GeoDataFrame
+from matplotlib.axes import Axes
 from matplotlib.cm import ScalarMappable
-from matplotlib.colors import Normalize
+from matplotlib.colors import (
+    Colormap,
+    LinearSegmentedColormap,
+    ListedColormap,
+    Normalize,
+)
+from matplotlib.transforms import CompositeGenericTransform
+from numpy.ma.core import MaskedArray
+from spatialdata.models import Image2DModel, SpatialElement
+from spatialdata.transformations.operations import get_transformation
+from spatialdata.transformations.transformations import Scale, Translation
+from spatialdata.transformations.transformations import Sequence as TransformSequence
+from xarray import DataArray
 
 from spatialdata_plot._logging import logger
 from spatialdata_plot.pl.render_params import Color, FigParams, ShapesRenderParams, _DsReduction
 from spatialdata_plot.pl.utils import (
-    _DS_REDUCTION_FUNCS,
-    _ax_show_and_transform,
-    _convert_alpha_to_datashader_range,
-    _create_image_from_datashader_result,
-    _datashader_aggregate_with_function,
-    _datashader_map_aggregate_to_color,
-    _datshader_get_how_kw_for_spread,
-    _hex_no_alpha,
+    _fast_extent,
     _make_continuous_mappable,
+    to_hex,
 )
 
 # ---------------------------------------------------------------------------
@@ -244,7 +258,7 @@ def _ds_shade_continuous(
     if spread_px is not None:
         # markers overlay (don't accumulate): spread with "max" so overlapping dots keep the true
         # value range instead of summing and inflating the colorbar (see as_points).
-        spread_how = "max" if uniform_alpha else _datshader_get_how_kw_for_spread(ds_reduction)
+        spread_how = "max" if uniform_alpha else _datashader_get_how_kw_for_spread(ds_reduction)
         agg = ds.tf.spread(agg, px=spread_px, how=spread_how)
         reduction_bounds = (agg.min(), agg.max())
 
@@ -517,3 +531,419 @@ def _build_ds_colorbar(
     vmin = reduction_bounds[0].values if norm.vmin is None else norm.vmin
     vmax = reduction_bounds[1].values if norm.vmax is None else norm.vmax
     return _make_continuous_mappable(vmin, vmax, cmap)
+
+
+# ---------------------------------------------------------------------------
+# Datashader reduction constants
+# ---------------------------------------------------------------------------
+
+
+_DS_REDUCTION_FUNCS: dict[str, Any] = {
+    "sum": ds.sum,
+    "mean": ds.mean,
+    "any": ds.any,
+    "count": ds.count,
+    "std": ds.std,
+    "var": ds.var,
+    "max": ds.max,
+    "min": ds.min,
+}
+
+
+# ---------------------------------------------------------------------------
+# Color / alpha helpers
+# ---------------------------------------------------------------------------
+
+
+def _hex_no_alpha(hex: str) -> str:
+    """
+    Return a hex color string without an alpha component.
+
+    Parameters
+    ----------
+    hex : str
+        The input hex color string. Must be in one of the following formats:
+        - "#RRGGBB": a hex color without an alpha channel.
+        - "#RRGGBBAA": a hex color with an alpha channel that will be removed.
+
+    Returns
+    -------
+    str
+        The hex color string in "#RRGGBB" format.
+    """
+    if not isinstance(hex, str):
+        raise TypeError("Input must be a string")
+    if not hex.startswith("#"):
+        raise ValueError("Invalid hex color: must start with '#'")
+
+    hex_digits = hex[1:]
+    length = len(hex_digits)
+
+    if length == 6:
+        if not all(c in "0123456789abcdefABCDEF" for c in hex_digits):
+            raise ValueError("Invalid hex color: contains non-hex characters")
+        return hex  # Already in #RRGGBB format.
+
+    if length == 8:
+        if not all(c in "0123456789abcdefABCDEF" for c in hex_digits):
+            raise ValueError("Invalid hex color: contains non-hex characters")
+        # Return only the first 6 characters, stripping the alpha.
+        return "#" + hex_digits[:6]
+
+    raise ValueError("Invalid hex color length: must be either '#RRGGBB' or '#RRGGBBAA'")
+
+
+def _convert_alpha_to_datashader_range(alpha: float) -> float:
+    """Convert alpha from the range [0, 1] to the range [0, 255] used in datashader."""
+    # prevent a value of 255, bc that led to fully colored test plots instead of just colored points/shapes
+    return min([254, alpha * 255])
+
+
+# ---------------------------------------------------------------------------
+# Canvas geometry helpers
+# ---------------------------------------------------------------------------
+
+
+def _ax_show_and_transform(
+    array: MaskedArray[tuple[int, ...], Any] | npt.NDArray[Any],
+    trans_data: CompositeGenericTransform,
+    ax: Axes,
+    alpha: float | None = None,
+    cmap: ListedColormap | LinearSegmentedColormap | None = None,
+    zorder: int = 0,
+    norm: Normalize | None = None,
+    interpolation: str | None = None,
+) -> matplotlib.image.AxesImage:
+    # ``extent`` uses mpl's pixel-grid convention; world placement happens via
+    # ``set_transform(trans_data)`` afterwards.
+    image_extent = (-0.5, array.shape[1] - 0.5, array.shape[0] - 0.5, -0.5)
+    # ``alpha`` is applied only when no cmap is set, so RGBA arrays already
+    # carrying per-pixel alpha (e.g. datashader output) are not double-attenuated.
+    imshow_kwargs: dict[str, Any] = {"zorder": zorder, "extent": image_extent, "norm": norm}
+    if not cmap and alpha is not None:
+        imshow_kwargs["alpha"] = alpha
+    else:
+        imshow_kwargs["cmap"] = cmap
+    if interpolation is not None:
+        imshow_kwargs["interpolation"] = interpolation
+    im = ax.imshow(array, **imshow_kwargs)
+    im.set_transform(trans_data)
+    return im
+
+
+def _compute_datashader_canvas_params(
+    x_ext: list[Any],
+    y_ext: list[Any],
+    fig_params: FigParams,
+) -> tuple[Any, Any, list[Any], list[Any], Any]:
+    """Compute datashader canvas dimensions from spatial extents.
+
+    Shared logic used by both the dask-based and pandas-based entry points.
+    """
+    # Compute canvas size in pixels, capped at the figure's display resolution.
+    # Using np.max ensures the canvas never exceeds display pixels on either axis,
+    # preventing pixel-based operations (spread, line_width) from being downscaled
+    # to sub-pixel size when the data aspect ratio differs from the figure's.
+    plot_width = x_ext[1] - x_ext[0]
+    plot_height = y_ext[1] - y_ext[0]
+    plot_width_px = int(round(fig_params.fig.get_size_inches()[0] * fig_params.fig.dpi))
+    plot_height_px = int(round(fig_params.fig.get_size_inches()[1] * fig_params.fig.dpi))
+    factor: float
+    factor = np.max([plot_width / plot_width_px, plot_height / plot_height_px])
+    plot_width = int(np.round(plot_width / factor))
+    plot_height = int(np.round(plot_height / factor))
+
+    return plot_width, plot_height, x_ext, y_ext, factor
+
+
+def _get_extent_and_range_for_datashader_canvas(
+    spatial_element: SpatialElement,
+    coordinate_system: str,
+    fig_params: FigParams,
+) -> tuple[Any, Any, list[Any], list[Any], Any]:
+    extent = _fast_extent(spatial_element, coordinate_system)
+    x_ext = [float(extent["x"][0]), float(extent["x"][1])]
+    y_ext = [float(extent["y"][0]), float(extent["y"][1])]
+    return _compute_datashader_canvas_params(x_ext, y_ext, fig_params)
+
+
+def _datashader_canvas_from_dataframe(
+    df: pd.DataFrame,
+    fig_params: FigParams,
+) -> tuple[Any, Any, list[Any], list[Any], Any]:
+    """Compute datashader canvas params directly from a pandas DataFrame.
+
+    Avoids the overhead of ``get_extent()`` (which requires a dask-backed
+    SpatialElement) by reading min/max from the already-materialised data.
+    """
+    if len(df) == 0:
+        # Empty input (e.g., a bounding_box_query with no overlap) — caller
+        # should short-circuit; return zero-sized canvas params as a sentinel.
+        return 0, 0, [0.0, 0.0], [0.0, 0.0], 1.0
+    x_ext = [float(df["x"].min()), float(df["x"].max())]
+    y_ext = [float(df["y"].min()), float(df["y"].max())]
+    return _compute_datashader_canvas_params(x_ext, y_ext, fig_params)
+
+
+def _prepare_transformation(
+    element: DataArray | GeoDataFrame | dask.dataframe.core.DataFrame,
+    coordinate_system: str,
+    ax: Axes | None = None,
+) -> tuple[
+    matplotlib.transforms.Affine2D,
+    matplotlib.transforms.CompositeGenericTransform | None,
+]:
+    trans = get_transformation(element, get_all=True)[coordinate_system]
+    affine_trans = trans.to_affine_matrix(input_axes=("x", "y"), output_axes=("x", "y"))
+    trans = mtransforms.Affine2D(matrix=affine_trans)
+    trans_data = trans + ax.transData if ax is not None else None
+
+    return trans, trans_data
+
+
+def _create_image_from_datashader_result(
+    ds_result: ds.transfer_functions.Image | np.ndarray[Any, np.dtype[np.uint8]],
+    factor: float,
+    ax: Axes,
+    x_min: float = 0.0,
+    y_min: float = 0.0,
+) -> tuple[MaskedArray[tuple[int, ...], Any], matplotlib.transforms.Transform]:
+    # create SpatialImage from datashader output to get it back to original size
+    rgba_image_data = ds_result.copy() if isinstance(ds_result, np.ndarray) else ds_result.to_numpy().base
+    rgba_image_data = np.transpose(rgba_image_data, (2, 0, 1))
+    transformation: Scale | TransformSequence = Scale([1, factor, factor], ("c", "y", "x"))
+    if x_min != 0.0 or y_min != 0.0:
+        # Canvas pixel (0, 0) corresponds to world (x_min, y_min). Without this
+        # translation the rgba would render at the world origin instead of at
+        # the element's actual position.
+        transformation = TransformSequence([transformation, Translation([x_min, y_min], ("x", "y"))])
+    rgba_image = Image2DModel.parse(
+        rgba_image_data,
+        dims=("c", "y", "x"),
+        transformations={"global": transformation},
+    )
+
+    _, trans_data = _prepare_transformation(rgba_image, "global", ax)
+
+    rgba_image = np.transpose(rgba_image.data.compute(), (1, 2, 0))  # type: ignore[attr-defined]
+    rgba_image = ma.masked_array(rgba_image)  # type conversion for mypy
+
+    return rgba_image, trans_data
+
+
+# ---------------------------------------------------------------------------
+# Aggregation / shading helpers
+# ---------------------------------------------------------------------------
+
+
+def _datashader_aggregate_with_function(
+    reduction: _DsReduction | None,
+    cvs: Canvas,
+    spatial_element: GeoDataFrame | dask.dataframe.core.DataFrame,
+    col_for_color: str | None,
+    element_type: Literal["points", "shapes"],
+) -> DataArray:
+    """
+    When shapes or points are colored by a continuous value during rendering with datashader.
+
+    This function performs the aggregation using the user-specified reduction method.
+
+    Parameters
+    ----------
+    reduction: String specifying the datashader reduction method to be used.
+        If None, "sum" is used as default.
+    cvs: Canvas object previously created with ds.Canvas()
+    spatial_element: geo or dask dataframe with the shapes or points to render
+    col_for_color: name of the column containing the values by which to color
+    element_type: tells us if this function is called from _render_shapes() or _render_points()
+    """
+    if reduction is None:
+        reduction = "sum"
+
+    try:
+        reduction_function = _DS_REDUCTION_FUNCS[reduction](column=col_for_color)
+    except KeyError as e:
+        raise ValueError(
+            f"Reduction '{reduction}' is not supported. Please use one of: {', '.join(_DS_REDUCTION_FUNCS.keys())}."
+        ) from e
+
+    element_function_map = {
+        "points": cvs.points,
+        "shapes": cvs.polygons,
+    }
+
+    try:
+        element_function = element_function_map[element_type]
+    except KeyError as e:
+        raise ValueError(f"Element type '{element_type}' is not supported. Use 'points' or 'shapes'.") from e
+
+    if element_type == "points":
+        points_aggregate = element_function(spatial_element, "x", "y", agg=reduction_function)
+        if reduction == "any":
+            # replace False/True by nan/1
+            points_aggregate = points_aggregate.astype(int)
+            points_aggregate = points_aggregate.where(points_aggregate > 0)
+        return points_aggregate
+
+    # is shapes
+    return element_function(spatial_element, geometry="geometry", agg=reduction_function)
+
+
+def _datashader_get_how_kw_for_spread(
+    reduction: _DsReduction | None,
+) -> str:
+    # Get the best input for the how argument of ds.tf.spread(), needed for numerical values
+    reduction = reduction or "sum"
+
+    reduction_to_how_map = {
+        "sum": "add",
+        "mean": "source",
+        "any": "source",
+        "count": "add",
+        "std": "source",
+        "var": "source",
+        "max": "max",
+        "min": "min",
+    }
+
+    if reduction not in reduction_to_how_map:
+        raise ValueError(
+            f"Reduction {reduction} is not supported, please use one of the following: sum, mean, any, count"
+            ", std, var, max, min."
+        )
+
+    return reduction_to_how_map[reduction]
+
+
+def _apply_cmap_alpha_to_datashader_result(
+    result: Any,
+    agg: DataArray,
+    cmap: str | list[str] | Colormap,
+    span: list[float] | tuple[float, float] | None,
+) -> Any:
+    """Apply the colormap's alpha channel to a datashader RGBA result.
+
+    Datashader ignores the per-entry alpha channel of matplotlib colormaps,
+    so pixels that the cmap marks as transparent (alpha=0) are rendered
+    opaque.  This function post-processes the shaded RGBA output to restore
+    the cmap's intended transparency.  See :issue:`376`.
+    """
+    if not isinstance(cmap, Colormap):
+        return result
+
+    # Quick check: does this cmap have any transparent entries?
+    test_vals = np.linspace(0, 1, min(cmap.N, 256))
+    cmap_alphas = cmap(test_vals)[:, 3]
+    if np.all(cmap_alphas >= 1.0):
+        return result
+
+    # Get or ensure we have an (H, W, 4) uint8 array
+    if hasattr(result, "values"):
+        # datashader Image — uint32 packed, convert via to_numpy()
+        rgba = result.to_numpy().base
+        if rgba is None:
+            return result
+    else:
+        rgba = result
+
+    if rgba.ndim != 3 or rgba.shape[2] != 4:
+        return result
+
+    # Normalise aggregate values to [0, 1] using the same span datashader used
+    agg_vals = agg.values.astype(np.float64)
+    valid = np.isfinite(agg_vals)
+    if not valid.any():
+        return result
+
+    if span is not None:
+        lo, hi = float(span[0]), float(span[1])
+    else:
+        lo = float(np.nanmin(agg_vals))
+        hi = float(np.nanmax(agg_vals))
+
+    if hi <= lo or not np.isfinite(lo) or not np.isfinite(hi):
+        return result
+
+    normed = np.clip((agg_vals - lo) / (hi - lo), 0.0, 1.0)
+
+    # Look up cmap alpha for each pixel
+    desired_alpha = cmap(normed)[:, :, 3]
+
+    # Zero out pixels where the cmap wants transparency
+    transparent = valid & (desired_alpha < 1.0)
+    if transparent.any():
+        # Scale the existing alpha by the cmap's alpha
+        rgba[transparent, 3] = (rgba[transparent, 3].astype(np.float32) * desired_alpha[transparent]).astype(np.uint8)
+
+    return result
+
+
+def _datashader_map_aggregate_to_color(
+    agg: DataArray,
+    cmap: str | list[str] | ListedColormap,
+    color_key: list[str] | dict[str, str] | None = None,
+    min_alpha: float = 40,
+    span: None | list[float] = None,
+    clip: bool = True,
+    how: str = "linear",
+) -> ds.tf.Image | np.ndarray[Any, np.dtype[np.uint8]]:
+    """ds.tf.shade() part, ensuring correct clipping behavior.
+
+    If necessary (norm.clip=False), split shading in 3 parts and in the end, stack results.
+    This ensures the correct clipping behavior, because else datashader would always automatically clip.
+
+    ``how`` controls the count-to-color mapping passed to :func:`datashader.transfer_functions.shade`
+    (``"linear"`` by default; ``"log"``/``"cbrt"``/``"eq_hist"`` compress dynamic range). The split-shade
+    branch used for ``norm.clip=False`` always uses ``"linear"`` since per-segment shading would otherwise
+    interact poorly with rank-based mappings.
+    """
+    if not clip and isinstance(cmap, Colormap) and span is not None:
+        # in case we use datashader together with a Normalize object where clip=False
+        # why we need this is documented in https://github.com/scverse/spatialdata-plot/issues/372
+        agg_in = agg.where((agg >= span[0]) & (agg <= span[1]))
+        img_in = ds.tf.shade(
+            agg_in,
+            cmap=cmap,
+            span=(span[0], span[1]),
+            how="linear",
+            color_key=color_key,
+            min_alpha=min_alpha,
+        )
+
+        agg_under = agg.where(agg < span[0])
+        img_under = ds.tf.shade(
+            agg_under,
+            cmap=[to_hex(cmap.get_under())[:7]],
+            min_alpha=min_alpha,
+            color_key=color_key,
+        )
+
+        agg_over = agg.where(agg > span[1])
+        img_over = ds.tf.shade(
+            agg_over,
+            cmap=[to_hex(cmap.get_over())[:7]],
+            min_alpha=min_alpha,
+            color_key=color_key,
+        )
+
+        # stack the 3 arrays manually: go from under, through in to over and always overlay the values where alpha=0
+        stack = img_under.to_numpy().base
+        if stack is None:
+            stack = img_in.to_numpy().base
+        else:
+            stack[stack[:, :, 3] == 0] = img_in.to_numpy().base[stack[:, :, 3] == 0]
+        img_over = img_over.to_numpy().base
+        if img_over is not None:
+            stack[stack[:, :, 3] == 0] = img_over[stack[:, :, 3] == 0]
+
+        return _apply_cmap_alpha_to_datashader_result(stack, agg, cmap, span)
+
+    result = ds.tf.shade(
+        agg,
+        cmap=cmap,
+        color_key=color_key,
+        min_alpha=min_alpha,
+        span=span,
+        how=how,
+    )
+    return _apply_cmap_alpha_to_datashader_result(result, agg, cmap, span)
