+++ src/spatialdata_plot/pl/render.py
@@ -34,11 +34,16 @@ from xarray import DataTree
 from spatialdata_plot._logging import _log_context, logger
 from spatialdata_plot.pl._datashader import (
     _apply_ds_norm,
+    _ax_show_and_transform,
     _build_color_key,
     _build_ds_colorbar,
+    _datashader_canvas_from_dataframe,
     _ds_aggregate,
     _ds_shade_categorical,
     _ds_shade_continuous,
+    _get_extent_and_range_for_datashader_canvas,
+    _hex_no_alpha,
+    _prepare_transformation,
     _render_ds_image,
     _render_ds_outlines,
 )
@@ -65,16 +70,12 @@ from spatialdata_plot.pl.render_params import (
 from spatialdata_plot.pl.utils import (
     _align_outline_vector_to_length,
     _apply_mask_to_outline_vectors,
-    _ax_show_and_transform,
     _check_obs_var_shadow,
     _color_vector_to_rgba,
-    _datashader_canvas_from_dataframe,
     _decorate_axs,
     _fast_extent,
     _get_colors_for_categorical_obs,
-    _get_extent_and_range_for_datashader_canvas,
     _get_linear_colormap,
-    _hex_no_alpha,
     _join_table_for_element,
     _make_continuous_mappable,
     _map_color_seg,
@@ -83,7 +84,6 @@ from spatialdata_plot.pl.utils import (
     _multiscale_to_spatial_image,
     _pixel_to_coord,
     _prepare_cmap_norm,
-    _prepare_transformation,
     _rasterize_if_necessary,
     _rasterize_if_necessary_datashader,
     _set_color_source_vec,
