+++ tests/pl/test_render_shapes.py
@@ -442,16 +442,14 @@ class TestShapes(PlotTester, metaclass=PlotTesterMeta):
     def test_plot_shapes_unannotated_by_table_render_with_na_color(self, sdata_blobs_shapes_annotated: SpatialData):
         # Regression for #710: blobs_polygons instance 0 has no row in the table (instance_id starts
         # at 1), so coloring by a table column must render it with na_color, not drop it.
-        sdata_blobs_shapes_annotated.pl.render_shapes(
-            "blobs_polygons", color="channel_0_sum", na_color="red"
-        ).pl.show()
+        sdata_blobs_shapes_annotated.pl.render_shapes("blobs_polygons", color="channel_0_sum", na_color="red").pl.show()
 
-    def test_plot_shapes_unannotated_by_table_hidden_with_na_color_none(self, sdata_blobs_shapes_annotated: SpatialData):
+    def test_plot_shapes_unannotated_by_table_hidden_with_na_color_none(
+        self, sdata_blobs_shapes_annotated: SpatialData
+    ):
         # Counterpart to the test above: na_color=None makes the unannotated polygon (instance 0)
         # transparent, opting back into hiding it.
-        sdata_blobs_shapes_annotated.pl.render_shapes(
-            "blobs_polygons", color="channel_0_sum", na_color=None
-        ).pl.show()
+        sdata_blobs_shapes_annotated.pl.render_shapes("blobs_polygons", color="channel_0_sum", na_color=None).pl.show()
 
     def test_plot_can_plot_shapes_after_spatial_query(self, sdata_blobs: SpatialData):
         # subset to only shapes, should be unnecessary after rasterizeation of multiscale images is included
@@ -1226,7 +1224,7 @@ def test_groups_filtering_preserves_transformation(sdata_blobs: SpatialData):
     re-assign to sdata_filt -> GeoDataFrame re-wrap — then asserts that
     ``_prepare_transformation`` can still retrieve the correct transformation.
     """
-    from spatialdata_plot.pl.utils import _prepare_transformation
+    from spatialdata_plot.pl._datashader import _prepare_transformation
 
     scale_factor = 2.5
     cs = "not_global"
@@ -1753,9 +1751,7 @@ def test_render_shapes_as_points_applies_non_identity_transform(sdata_blobs: Spa
 
     set_transformation(sdata_blobs["blobs_circles"], Scale([3.0, 5.0], axes=("x", "y")), "scaled")
     fig, ax = plt.subplots()
-    sdata_blobs.pl.render_shapes("blobs_circles", as_points=True, size=50).pl.show(
-        ax=ax, coordinate_systems="scaled"
-    )
+    sdata_blobs.pl.render_shapes("blobs_circles", as_points=True, size=50).pl.show(ax=ax, coordinate_systems="scaled")
     coll = ax.collections[0]
     dots = coll.get_offset_transform().transform(np.asarray(coll.get_offsets()))
     cs = sd.get_centroids(sdata_blobs["blobs_circles"], coordinate_system="scaled").compute()[["x", "y"]].to_numpy()
